Remove time_values and geo_values arguments from plot_line
A covidcast_signal df is just a df; we can trust the user to be able to
filter it however they desire. That's the typical interface for plotting
functions, including ggplot2, which just expect you to wrangle your data
frame as desired and then provide it. This lets the user be flexible
about what they want to select and how.