Use rlang::abort and rlang::warn for recoverable/loggable errors
warning() raises a warning condition that can be handled via
withCallingHandlers, e.g. if you want to log warnings during program
execution. But the warnings have no class attached, so the only way to
catch specific warnings would be pattern-matching the message.

Instead, use rlang's convenience functions for raising warnings and
errors with structured types. Attach metadata to the warnings so a
logger can use the information as needed.

Fixes #63.