Introduce as.covidcast_signal generic function
This function handles everything needed to turn a data frame into a
covidcast_signal, including attaching necessary columns and metadata.

Along the way, fix the wrangling functions to preserve metadata and
attach it. Also, ensure that the basic metadata always includes source
and signal -- if it doesn't, when signals are aggregated it will be
impossible to tell which row of the metadata corresponds to which
signal!