Correct inclusion criteria for county plots
They were confused by state FIPS codes. fips_to_abbr is a footgun and
should be revised later, but this fixes the immediate issue and adds a
test. (Counties from other states were included in the Alaska plot, with
a warning for duplicate matches, because FIPS "2" matched many states
with "2" in their state FIPS.)